Передалал компонент StateBtn на использование UI
