+++ src/constants/ToDo.ts
@@ -16,15 +16,3 @@ export type Task = {
     state: States;
     label: string;
 };
-
-/**
- * Цвета кнопок класс bootstrap
- */
-export enum ButtonColors {
-    red = "text-danger",
-    blue = "text-primary",
-    gray = "text-secondary",
-    green = "text-success",
-    yellow = "text-warning",
-    black = "text-dark",
-}
